Rebuild CLI to not depend on click, and drop .main()
We are now dep free!  Since all we want to do is pass our args through
to terraform (if any) this was a straight forward thing to do without
any library code.

If you don't specify any args then terraformpy just writes the tf.json
file and exits.

We also used to call .main() on the imported .tf.py file.  Given the
design we've landed on for how things are organized this isn't
necessary.