Collection "finalization" & Frozen objects
This allows for collections to be finalized before compilation.  The
main motivator for this we want to build a Security Group collection
that allows us to model rules in pure python and then generate a number
of individual rules right before things are compiled.

The Frozen objects allows you to reach into the `_values` attr of a
TFObject (that provides interpolation) through the magic attr getter
function until compilation time.

Closes #9